==> 4.pathway_expression_clustering.Rmd <==
---
title: "Pathway Expression Clustering"
author: "Dr Jacob Househam"
date: "21/06/2021"
output: html_document
---

```{r library, message=FALSE, warning=FALSE, include=FALSE}
# Specifically requires misgdbr 7.2.1 (not 7.4.2)
library(data.table);library(DESeq2);library(wesanderson);library(pheatmap);library(dplyr);library(analogue);library(GSVA);library(msigdbr)
options(datatable.fread.datatable=FALSE)
```

```{r setup}
ngroupcol= c(wes_palette("Darjeeling2",5)[2],wes_palette('Darjeeling1',5)[2],wes_palette('FantasticFox1',5)[c(5,2)])
names(ngroupcol) <- as.numeric(c(1:4))
colhall <- c('#007F00','#CC0000','purple','#0000E5','#D3D3D3');names(colhall) <- c('Immune','Oncogenic','Cellular Stress','Stromal','Other')
col_list <- list(MSI_Status=c(MSI='#01DADF',MSS='#FF9288'),
                 Group=ngroupcol,Class=colhall)
```

## Load MSigDB Hallmark pathways and slightly modify
```{r pathway_gather}
m_df = msigdbr(species = "Homo sapiens", category = "H")
Hallmark_cancer_list = split(x = m_df$entrez_gene, f = m_df$gs_name)
names(Hallmark_cancer_list) <- gsub('HALLMARK_(\\S+)','\\1',names(Hallmark_cancer_list))
# Remove spermatogenesis, myogenesis and pancreas_beta_cells
Hallmark_cancer_list$SPERMATOGENESIS <- NULL
Hallmark_cancer_list$MYOGENESIS <- NULL
Hallmark_cancer_list$PANCREAS_BETA_CELLS <- NULL
# Rename COMPLEMENT
Hallmark_cancer_list$COMPLEMENT_INNATE_IMMUNE_SYSTEM <- Hallmark_cancer_list$COMPLEMENT
Hallmark_cancer_list$COMPLEMENT <- NULL

# Include ISC signature and WNT_signalling
isc <- read.table('input_files/merlossuarez_2011_ISC_signature.txt')[,1]
geneinfo <- fread('input_files/complete_gene_info.txt.gz')
isc_entrez <- geneinfo[which(geneinfo$Symbol %in% isc),'Entrez']
isc_entrez <- isc_entrez[which(!is.na(isc_entrez))]
Hallmark_cancer_list$INTESTINAL_STEM_CELL <- isc_entrez
wnt_entrez <- read.table('input_files/wnt_signalling_entrez.txt')[,1]
Hallmark_cancer_list$WNT_SIGNALING <- wnt_entrez
```

## Load sample data and TPM expression and convert to entrez IDs
```{r load_sample_and_expression}
# Load sample data and filter for multi-tumour sample patients
passsam <- read.table('input_files/ListRNAPass.EPICC.txt')[,1]
tsam <- passsam[-grep('^C\\d+_E',passsam)]
allpats <- gsub('^(C\\d+)_\\S+','\\1',tsam);patients <- unique(allpats);numsam <- table(allpats)
tsam <- tsam[which(allpats %in% names(which(numsam>=5)))]
allpats <- gsub('^(C\\d+)_\\S+','\\1',tsam);patients <- unique(allpats);numsam <- table(allpats)
msipat <- c('C516','C518','C536','C548','C552')
normsam <- passsam[grep('^C\\d+_E',passsam)];allnormpats <- gsub('^(C\\d+)_\\S+','\\1',normsam)
normsam <- normsam[which(allnormpats %in% allpats)]

# Load tpm, convert to log(TPM+1) expression and convert to entrez gene ids
tmptpm <- readRDS('intermediates/All_EPICC_tpm.rds')
#tmptpm <- fread('~/Documents/EPICC/Data/Expression/ProcessedCounts/All_EPICC_tpm.txt')
tpm <- tmptpm
for(i in c(2:ncol(tmptpm))) { tpm[,i] <- log(tmptpm[,i]+1) }
geneinfoexp <- merge(tpm,geneinfo,by='GeneID')
# Remove NA entrez genes
geneinfoexp <- geneinfoexp[!is.na(geneinfoexp$Entrez),];row.names(geneinfoexp) <- c(1:nrow(geneinfoexp))
# For now if 1 ensembl ids maps to multiple entrez, choose the first entrez
geneinfoexp[grep(',',geneinfoexp$Entrez),'Entrez'] <- gsub('^(\\S+),\\S+','\\1',geneinfoexp[grep(',',geneinfoexp$Entrez),'Entrez'])
# Temp remove duplcates
geneinfoexp <- geneinfoexp[which(!duplicated(geneinfoexp$Entrez)),];row.names(geneinfoexp) <- c(1:nrow(geneinfoexp))
# Remove genes with 0 expression for all samples
entrezexp <- geneinfoexp
row.names(entrezexp) <- entrezexp$Entrez
norm_entrezexp <- entrezexp[,normsam]
entrezexp <- entrezexp[,tsam]

```


## Run ssGSEA and calculate mean enrichment and mean standard deviation per pathway
```{r get_mean_and_sd}
expmat <- as.data.frame(matrix(0L,nrow=length(names(Hallmark_cancer_list)),ncol=length(patients)))
row.names(expmat) <- names(Hallmark_cancer_list);colnames(expmat) <- patients
sdmat <- expmat
for(pat in patients) {
  patsam <- tsam[grep(pat,tsam)]
  pattpm <- entrezexp[,patsam]
  hallmark_gsva <- gsva(data.matrix(pattpm),method='ssgsea',
                        Hallmark_cancer_list,verbose=FALSE)
  expmat[[pat]] <- rowMeans(as.matrix(hallmark_gsva))
  sdmat[[pat]] <- apply(hallmark_gsva,1,function(x) sd(x))
}

# Additionally get the enrichment for the set of normal samples
norm_hallmark <- gsva(data.matrix(norm_entrezexp),method='ssgsea',
                      Hallmark_cancer_list,verbose=FALSE)
normexp <- rowMeans(as.matrix(norm_hallmark))
```

## Create pathdf summarising pathway info
```{r create_path_df}
pathdf <- data.frame(Mean_Var=rowMeans(as.matrix(sdmat)),Mean_Mean_Enrichment=rowMeans(as.matrix(log(expmat+1))),Mean_Norm=normexp)
row.names(pathdf) <- names(Hallmark_cancer_list)
pathdf <- pathdf[which(pathdf$Mean_Mean_Enrichment>0),]

# Load information of hallmark categories
info_hallmark <- fread('input_files/hallmark_info.txt')
# Made by PMID: 32483290, got classes by looking at their github - specifically:
# Line 21 of https://nbviewer.jupyter.org/github/cansysbio/HGSOC_TME_Heterogeneity/blob/master/Scripts/1/1_09.FactorsMap_Hallmark_NES_TreatmentNaive.ipynb for the pathways -> colours
# Line 287-288 of https://github.com/cansysbio/HGSOC_TME_Heterogeneity/blob/b697b8a6113d31196715f6ef581648237f8e9414/Scripts/4/4/scripts/corSignaturesGeneExprCNA.R for the colors -> classes

info_hallmark$Hallmark_name <- gsub('HALLMARK_(\\S+)','\\1',info_hallmark$Hallmark_name)
# Add categories for custom pathways and change name for COMPLEMENT
info_hallmark[nrow(info_hallmark)+1,] <- c('INTESTINAL_STEM_CELL','signalling','Oncogenic','Intestinal stem cell',10,length(isc_entrez))
info_hallmark[nrow(info_hallmark)+1,] <- c('WNT_SIGNALING','signalling','Oncogenic','IGenes related to Wnt-mediated signal transduction',10,length(wnt_entrez))
info_hallmark[which(info_hallmark$Hallmark_name=='COMPLEMENT'),'Hallmark_name'] <- 'COMPLEMENT_INNATE_IMMUNE_SYSTEM'

# Merge with pathdf
pathdf <- merge(pathdf,info_hallmark,by.x=0,by.y='Hallmark_name');row.names(pathdf) <- pathdf$Row.names
pathdf <- pathdf[,c('Class','Mean_Var','Mean_Mean_Enrichment','Mean_Norm')]
```

## Calculate distance matrices and cluster
```{r dist_and_cluster}
# First filter for useable pathways
sortexpmat <- expmat[row.names(pathdf),];sortsdmat <- sdmat[row.names(pathdf),]

# Then sort pathways by enrichment and sd of MSS tumours 
dexp <- dist(sortexpmat[,colnames(sortexpmat)[which(colnames(sortexpmat) %ni% msipat)]],method='euclidean')
dsd <-  dist(sortsdmat[,colnames(sortsdmat)[which(colnames(sortsdmat) %ni% msipat)]],method='euclidean')
dd <- fuse(dexp, dsd, weights=c(0.5,0.5))
clustord <- hclust(dd,method='complete')

# Now order groups 1-4 according to mean_mean_exp
tmpmemb <- cutree(clustord,k=4)
medstat <- boxplot(pathdf$Mean_Mean_Enrichment~tmpmemb,plot=F)$stats[3,]
ordgroup <- order(medstat,decreasing=T)
pathdf$Group <- rep('1',nrow(pathdf));pathdf[tmpmemb==ordgroup[2],'Group'] <- '2'
pathdf[tmpmemb==ordgroup[3],'Group'] <- '3';pathdf[tmpmemb==ordgroup[4],'Group'] <- '4'

pathdf[pathdf$Group=='1','Group'] <- '5';pathdf[pathdf$Group=='2','Group'] <- '1'
pathdf[pathdf$Group=='5','Group'] <- '2'
saveRDS(pathdf,'results/pathway_clustering.rds')

# Next sort MSS samples according to exp and sd, and add MSI samples onto the end
dsamexp <- dist(t(sortexpmat[,colnames(sortexpmat)[which(colnames(sortexpmat) %ni% msipat)]]),method='euclidean')
dsamsd <- dist(t(sortsdmat[,colnames(sortsdmat)[which(colnames(sortsdmat) %ni% msipat)]]),method='euclidean')
ddsam <- fuse(dsamexp, dsamsd, weights=c(0.5,0.5))
clustsam <- hclust(ddsam,method='complete');samord <- c(colnames(sortexpmat)[which(colnames(sortexpmat) %ni% msipat)][clustsam$order],colnames(sortexpmat)[which(colnames(sortexpmat) %in% msipat)])

# Finally order exp and sd mats by clusters and then mean variance
sortexpmat <- arrange(sortexpmat,pathdf$Group,desc(pathdf$Mean_Var))
sortexpmat <- sortexpmat[,samord]
sortsdmat <- arrange(sortsdmat,pathdf$Group,desc(pathdf$Mean_Var))
sortsdmat <- sortsdmat[,samord]
```

## Plot clustering results (Extended Data Figure 1 A-C)
```{r plot_clustering}
patdf <- data.frame(MSI_Status=ifelse(colnames(expmat) %in% msipat,'MSI','MSS'))
row.names(patdf) <- colnames(expmat)
explot <- scale_mat(sortexpmat,'column')
sdplot <- scale_mat(sortsdmat,'column')
mybreaks <- seq(floor(min(explot,sdplot)),ceiling(max(explot,sdplot)),by=0.05)

pdf('figures/EDfig1A.expression.pathway.pdf',height=8,width=4)
pheatmap(explot,show_rownames=F,cluster_rows=F,cluster_cols=F,fontsize_row=7,
         gaps_col=15,fontsize_col=10,border_color=NA,fontsize=8,legend=F,
         color=wes_palette("Zissou1", length(mybreaks)-1,type="continuous"),breaks=mybreaks,
         cellwidth=9,cellheight=9,main='Mean ssGSEA Score',
         annotation_col=patdf,annotation_row=pathdf,annotation_names_col=F,
         annotation_colors=col_list,annotation_legend=F)
dev.off()

pdf('figures/EDfig1A.standarddeviation.pathway.pdf',height=7.5,width=5)
pheatmap(sdplot,show_rownames=T,cluster_rows=F,cluster_cols=F,fontsize_row=7,
         gaps_col=15,fontsize_col=10,border_color=NA,fontsize=8,
         color=wes_palette("Zissou1", length(mybreaks)-1,type="continuous"),breaks=mybreaks,
         cellwidth=9,cellheight=9,main='ssGSEA Score Standard Deviation',
         annotation_col=patdf,annotation_legend=F,#annotation_row=pathdf,
         annotation_colors=col_list)
dev.off()

# Exploring Group
pdf('figures/EDfig1BC.exploring_groups.pathway.pdf',width=4,height=4)
groupdf <- pathdf
groupdf$Pathway <- row.names(groupdf)
par(mfrow=c(1,1),font=2,font.axis=2,font.lab=2,
    mar=c(4.5,4.5,1,1))
xx <- boxplot(Mean_Mean_Enrichment~Group,data=groupdf,col=scales::alpha(ngroupcol,0.5),frame=F,
        xlab='',ylab='',boxcol=ngroupcol,boxlwd=1.5,medcol=ngroupcol,medlwd=2.5,
        staplecol=ngroupcol,staplelwd=1.5,whiskcol=ngroupcol,whisklwd=1.5,cex.axis=1.25,
        outpch=16,outcex=0.5,outcol=scales::alpha(ngroupcol,0.5),ylim=c(0,0.6),las=1)
mtext(side=1,text='Group',line=3);mtext(side=2,text='Mean Mean Enrichment',line=3.2)
mtext(side=1,at=c(1:4),text=paste0('n=',as.numeric(table(groupdf$Group))),line=2)
boxplot(Mean_Var~Group,data=groupdf,col=scales::alpha(ngroupcol,0.5),frame=F,
        xlab='',ylab='',boxcol=ngroupcol,boxlwd=1.5,medcol=ngroupcol,medlwd=2.5,
        staplecol=ngroupcol,staplelwd=1.5,whiskcol=ngroupcol,whisklwd=1.5,cex.axis=1.25,
        outpch=16,outcex=0.5,outcol=scales::alpha(ngroupcol,0.5),las=1)
mtext(side=1,text='Group',line=3);mtext(side=2,text='Mean Standard Deviation',line=3.2)
mtext(side=1,at=c(1:4),text=paste0('n=',as.numeric(table(groupdf$Group))),line=2)
dev.off()
```

## Perform and plot fishers test on pathway groups vs classes (Figure ED1D)
```{r fisher_groups_vs_classes}
pval <- matrix(0L,nrow=length(unique(pathdf$Class)),ncol=4);colnames(pval) <- c(1:4)
row.names(pval) <- unique(pathdf$Class)
lbnd <- ubnd <- OR <- pval
for(cat in unique(pathdf$Class)) {
  for(j in as.character(c(1:4))) {
    mat <- rbind(c(length(which(pathdf$Class!=cat & pathdf$Group!=j)),
                   length(which(pathdf$Class!=cat & pathdf$Group==j))),
                 c(length(which(pathdf$Class==cat & pathdf$Group!=j)),
                   length(which(pathdf$Class==cat & pathdf$Group==j))))
    res <- fisher.test(mat)
    pval[cat,j] <- res$p.value;lbnd[cat,j] <- res$conf.int[1]
    ubnd[cat,j] <- res$conf.int[2];OR[cat,j] <- res$estimate
  }
}

pdf('figures/EDfig1D.fisher_tests_group_vs_class.pdf',height=8,width=6)
options(scipen=-2)
sigcats <- c('Group2:Oncogenic','Group4:Stromal')
upcats <- c(sigcats,'Group3:Immune')
par(mar=c(3,7.3,1.5,10),font=2,font.axis=2,font.lab=2)
plot(x=c(0,40),y=c(1,20),axes=F,xlab='',ylab='',col='white')
abline(v=1,lty=3,col='black',lwd=3)
ypos <- 20;ynames <- c();pvals_vec <- c()
cxweight=1.5;lnweight=5
for(j in as.character(c(1:4))) {
  for(cat in unique(pathdf$Class)) {
    upper <- ifelse(is.infinite(ubnd[cat,j]),40,min(ubnd[cat,j],40))
    points(OR[cat,j],ypos,pch=16,col=ngroupcol[j],cex=cxweight)
    segments(x0=lbnd[cat,j],x1=upper,y0=ypos,col=ngroupcol[j],lwd=lnweight)
    
    if(paste0('Group',j,':',cat) %in% upcats) {
      segments(x0=40,x1=46,y0=ypos,col=ngroupcol[j],lwd=lnweight,xpd=T,lty=3)
      if(paste0('Group',j,':',cat)=='Group4:Stromal') {
        text(x=38,y=ypos+0.4,labels=paste0('OR & Upper 95% CI = ',ubnd[cat,j]),cex=0.6,xpd=T,adj=0.5)
      } else {
        text(x=38,y=ypos+0.4,labels=paste0('Upper 95% CI = ',signif(ubnd[cat,j],3)),cex=0.6,xpd=T,adj=0.5)
      }
    }
    ypos <- ypos-1
    ynames <- c(ynames,cat)
    pvals_vec <- c(pvals_vec,pval[cat,j])
  }
}
mtext(side=2,at=c(20:1),text=ynames,las=2,cex=1.2,line=0.5)
mtext(side=2,at=c(17.5,12.5,7.5,2.5)+0.5,text=paste0('Group ',c(1:4)),cex=1.2,line=5,col=ngroupcol,adj=0.3)
mtext(side=4,at=c(20:1),text=signif(pvals_vec,2),las=2,cex=1.3,line=2.5)
mtext(side=4,at=c(20:1),text=signif(p.adjust(pvals_vec,method='fdr'),2),las=2,cex=1.3,line=7)
axis(side=1,font=2,font.axis=2,cex.axis=1.2)
mtext(side=1,text='Odds Ratio',line=2,cex=1.2)
mtext(side=4,at=c(21,21),text=c('p-value','FDR'),line=c(1.5,6.5),las=1,adj=0,cex=1.3)
dev.off()
```
