==> 9.pathway_phylogenetic_analysis.Rmd <==
---
title: "Run and plot pathway enrichment phylogenetic signal analysis"
author: "Dr Jacob Househam"
date: "19/05/2022"
output: html_document
---

```{r library, message=FALSE, warning=FALSE, include=FALSE}
# Specifically requires misgdbr 7.2.1 (not 7.4.2)
library(phytools);library(data.table);library(DESeq2);library(GSVA);library(msigdbr);library(matrixStats);library(org.Hs.eg.db);library(wesanderson);library(pheatmap);library(phangorn)
options(datatable.fread.datatable=FALSE)
```

```{r setup}
ngroupcol= c(wes_palette("Darjeeling2",5)[2],wes_palette('Darjeeling1',5)[2],wes_palette('FantasticFox1',5)[c(5,2)])
names(ngroupcol) <- as.numeric(c(1:4))
colhall <- c('#007F00','#CC0000','purple','#0000E5','#D3D3D3');names(colhall) <- c('Immune','Oncogenic','Cellular Stress','Stromal','Other')
col_list <- list(Group=ngroupcol,Class=colhall)
```

## Gather pathways for analysis
```{r pathway_gather}
m_df = msigdbr(species = "Homo sapiens", category = "H")
Hallmark_cancer_list = split(x = m_df$entrez_gene, f = m_df$gs_name)
names(Hallmark_cancer_list) <- gsub('HALLMARK_(\\S+)','\\1',names(Hallmark_cancer_list))
# Remove spermatogenesis, myogenesis and pancreas_beta_cells
Hallmark_cancer_list$SPERMATOGENESIS <- NULL
Hallmark_cancer_list$MYOGENESIS <- NULL
Hallmark_cancer_list$PANCREAS_BETA_CELLS <- NULL
# Rename COMPLEMENT
Hallmark_cancer_list$COMPLEMENT_INNATE_IMMUNE_SYSTEM <- Hallmark_cancer_list$COMPLEMENT
Hallmark_cancer_list$COMPLEMENT <- NULL

# Include ISC signature and WNT_signalling
isc <- read.table('input_files/merlossuarez_2011_ISC_signature.txt')[,1]
geneinfo <- fread('input_files/complete_gene_info.txt.gz',data.table=F)
isc_entrez <- geneinfo[which(geneinfo$Symbol %in% isc),'Entrez']
isc_entrez <- isc_entrez[which(!is.na(isc_entrez))]
Hallmark_cancer_list$INTESTINAL_STEM_CELL <- isc_entrez
wnt_entrez <- read.table('input_files/wnt_signalling_entrez.txt')[,1]
Hallmark_cancer_list$WNT_SIGNALING <- wnt_entrez
```

## Load sample data, TPM and pathway clustering
```{r pathway_gather}
treelist <- readRDS('intermediates/pgls.trees.rds')
tinfo <- readRDS('intermediates/pgls.infotrees.rds')

# Load tpm, convert to log(TPM+1) expression and convert to entrez gene ids
tmptpm <- readRDS('intermediates/All_EPICC_tpm.rds');tpm <- tmptpm
for(i in c(2:ncol(tmptpm))) { tpm[,i] <- log(tmptpm[,i]+1) }
geneinfoexp <- merge(tpm,geneinfo,by='GeneID')
# Remove NA entrez genes
geneinfoexp <- geneinfoexp[!is.na(geneinfoexp$Entrez),];row.names(geneinfoexp) <- c(1:nrow(geneinfoexp))
# For now if 1 ensembl ids maps to multiple entrez, choose the first entrez
geneinfoexp[grep(',',geneinfoexp$Entrez),'Entrez'] <- gsub('^(\\S+),\\S+','\\1',geneinfoexp[grep(',',geneinfoexp$Entrez),'Entrez'])
# Temp remove duplcates
geneinfoexp <- geneinfoexp[which(!duplicated(geneinfoexp$Entrez)),];row.names(geneinfoexp) <- c(1:nrow(geneinfoexp))
# Remove genes with 0 expression for all samples
entrezexp <- geneinfoexp;row.names(entrezexp) <- entrezexp$Entrez
entrezexp <- entrezexp[,tinfo$samples]

# Input pathway clustering data
pathdf <- readRDS('results/pathway_clustering.rds')

```

## For each tumour get the phylogenetic signal (lambda) for all pathways for the 100 trees of varying branch length
```{r get_phylogenetic_signal}
pathway_lambda_list <- pathway_lambda_enrich <- list()
for(pat in tinfo$patients) {
  trimtree <- treelist[[pat]][[1]]
  treesam <- paste0(pat,"_",trimtree$tip.label)
  
  print(paste0(pat,': ',length(treesam),' matched samples x ',nrow(pathdf),' pathways'))
  pattpm <- entrezexp[which(apply(entrezexp[,treesam],1,sd)!=0),treesam]
  
  hallmark_gsva <- gsva(data.matrix(pattpm),method='ssgsea',
                        Hallmark_cancer_list,verbose=FALSE)
  try <- as.data.frame(t(hallmark_gsva[row.names(pathdf),treesam]))
  row.names(try) <- gsub('C\\d+_(\\S+)','\\1',row.names(try))
  
  siglist <- list()
  for(i in c(1:length(treelist[[pat]]))) {
    trimtree <- treelist[[pat]][[as.character(i)]]
    
    sigphylo <- data.frame(Pathway=colnames(try),Lambda=0,Lpval=0)
    for(j in c(1:ncol(try))) {
      curgen <- try[,j];names(curgen) <- row.names(try)
      
      res <- phylosig(trimtree,curgen,method='lambda',test=T)
      sigphylo[j,'Lambda'] <- res$lambda
      sigphylo[j,'Lpval'] <- res$P
    }
    siglist[[as.character(i)]] <- sigphylo
  }
  
  lam <- do.call(cbind, lapply(siglist,function(x) { x$Lambda }))
  lam_pval <- do.call(cbind, lapply(siglist,function(x) { x$Lpval }))
  resdf <- data.frame(MedLambda=rowMedians(lam),
                      LamPval=rowMedians(lam_pval))
  row.names(resdf) <- colnames(try)
  
  pathway_lambda_enrich[[pat]] <- try
  pathway_lambda_list[[pat]] <- resdf
}

saveRDS(pathway_lambda_enrich,file='results/pathway_lambda_enrichment.rds')
saveRDS(pathway_lambda_list,file='results/pathway_lambda_data.rds')
```

## Analyse and plot the results of the phylogenetic pathway enrichment analysis
### Figure 1J and Extended Data Figure 4
```{r plot_phylogenetic_pathway_results}
reclam <- matrix(0L,nrow=nrow(pathdf),ncol=length(tinfo$patients))
colnames(reclam) <- tinfo$patients;row.names(reclam) <- row.names(pathdf)
recstars <- recpval <- reclam
pdf('figures/fig1J&ED4.lambda_analysis_pathways.new.pdf',height=7,width=10)
for(pat in tinfo$patients) {
  resdf <- pathway_lambda_list[[pat]]
  try <- pathway_lambda_enrich[[pat]]
  
  combdf <- merge(resdf,pathdf,by=0);row.names(combdf) <- combdf$Row.names
  combdf$MedPval <- log(combdf$LamPval,base=10)*-1
  reclam[row.names(combdf),pat] <- combdf$MedLambda
  recstars[row.names(combdf),pat] <- getstars_vec(combdf$LamPval)
  recpval[row.names(combdf),pat] <- combdf$MedPval
  
  trimtree <- treelist[[pat]][[1]]
  
  regions <- gsub('^(\\S)\\d+_\\S+','\\1',trimtree$tip.label)
  options(scipen = -1);par(mar=c(0,0,2,0),xpd=T)
  plot.phylo(trimtree,type="phylogram",align.tip.label = T,edge.width=3,
             font=2,cex=0.9,tip.color=regcol[c(regions,'Root')],label.offset=20)
  title(main=pat)
  explot <- try
  
  # Change pathway names to numbers for editing (font size) reasons
  colnames(explot) <- row.names(combdf) <- c(1:ncol(explot))
  
  explot <- scale_mat(explot,'column')
  explot <- explot[c(nrow(explot):1),order(combdf$MedPval,decreasing=T)]
  mybreaks <- seq(floor(min(explot)),ceiling(max(explot)),by=0.05)
  pheatmap(explot,show_rownames = F,show_colnames=T,cluster_rows=F,cluster_cols=F,treeheight_col=0,
           border_color=NA,fontsize = 6,fontsize_col = 6,
           color=wes_palette("Zissou1", length(mybreaks)-1, type = "continuous"),breaks=mybreaks,
           annotation_col=combdf[,c('MedLambda','MedPval','Class','Group')],
           annotation_colors=col_list)
}
dev.off()

pathdf$Num_Sig <- rowSums(recstars!='')
saveRDS(pathdf,'results/reccurent_phylogenetic_pathways.rds')
```

## Plot phylogenetic recurrence of pathways
### Figure 1K
```{r plot_rec_phylo}
geneanot <- pathdf[row.names(reclam),c('Num_Sig','Group','Class')]
sortreclam <- reclam[order(geneanot$Num_Sig,geneanot$Group,decreasing=T),]
sortrecstars <- recstars[order(geneanot$Num_Sig,geneanot$Group,decreasing=T),]
sortrecpval <- recpval[order(geneanot$Num_Sig,geneanot$Group,decreasing=T),]

pdf('figures/fig1K.recurrent_pathways_lambda.pdf')
mybreaks <- seq(0,ceiling(max(sortrecpval)),by=0.01)
pheatmap(sortrecpval,show_rownames=T,show_colnames=T,cluster_rows=F,cluster_cols=T,
         border_color=NA,fontsize = 6,fontsize_col=9,fontsize_row=5,
         color=wes_palette("Zissou1", length(mybreaks)-1, type = "continuous"),breaks=mybreaks,
         display_numbers=sortrecstars,fontsize_number=6,
         annotation_row=geneanot,
         annotation_colors=col_list)
dev.off()

tmpgeneanot <- pathdf[row.names(reclam),c('Num_Sig','Group','Class')];row.names(tmpgeneanot) <- c(1:nrow(tmpgeneanot))
tmpreclam <- reclam;tmprecstars <- recstars;tmprecpval <- recpval
row.names(tmpreclam) <- row.names(tmprecstars) <- row.names(tmprecpval) <- c(1:nrow(tmpreclam))
sortreclam <- tmpreclam[order(tmpgeneanot$Num_Sig,tmpgeneanot$Group,decreasing=T),]
sortrecstars <- tmprecstars[order(tmpgeneanot$Num_Sig,tmpgeneanot$Group,decreasing=T),]
sortrecpval <- tmprecpval[order(tmpgeneanot$Num_Sig,tmpgeneanot$Group,decreasing=T),]

pdf('figures/fig1K.recurrent_pathways_lambda.alternative.pdf',height=8,width=3.5)
mybreaks <- seq(0,ceiling(max(sortrecpval)),by=0.01)
pheatmap(sortrecpval,show_rownames=T,show_colnames=T,cluster_rows=F,cluster_cols=T,legend = F,
         border_color=NA,fontsize=15,fontsize_col=20,fontsize_row=10,treeheight_col = 15,
         color=wes_palette("Zissou1", length(mybreaks)-1, type = "continuous"),breaks=mybreaks,
         display_numbers=sortrecstars,fontsize_number=20)
dev.off()
```
